#27: Changing FunctionalitiesManager in an abstract class commit does not compile
