+++ source/clients/admin/microkernel_architecture/microkernel_modules/functionalities_manager/FunctionalitiesManager.cpp
@@ -1,0 +1,22 @@
+#include "FunctionalitiesManager.hpp"
+
+//
+// Constructor : does nothing
+//
+FunctionalitiesManager::FunctionalitiesManager()
+{
+}
+
+//
+// Destructor : does nothing
+//
+FunctionalitiesManager::~FunctionalitiesManager()
+{
+}
+
+//
+// instanciate the various functionalities
+//
+void FunctionalitiesManager::_setFcts()
+{
+}
