+++ source/clients/common/microkernel_architecture/microkernel_modules/network_manager/NetworkClientManager.hpp
@@ -3,11 +3,10 @@
 
 # include <QSslSocket>
 # include "INetworkClientManager.hpp"
+# include "AFunctionality.hpp"
 
-class NetworkClientManager : public QObject, public INetworkClientManager
+class NetworkClientManager : public AFunctionality, public INetworkClientManager
 {
-  Q_OBJECT
-
 private:
     static const QString _ENCRYPTION_CERTIFICATE_FILE;
     static const QSsl::SslProtocol _DEFAULT_PROTOCOL;
@@ -16,7 +15,7 @@ private:
 
 public:
     explicit NetworkClientManager(QObject *parent = 0);
-    ~NetworkClientManager();
+    virtual ~NetworkClientManager();
 
     void startConnection(const QString &ip, quint16 port);
     void disconnect();
