+++ source/clients/user/microkernel_architecture/internal_fcts_modules/gui/gui_controller/GUIController.hpp
@@ -2,17 +2,16 @@
 # define GUI_CONTROLLER_HPP_
 
 # include <QQmlApplicationEngine>
+# include "AFunctionality.hpp"
 
-class GUIController : public QObject
+class GUIController : public AFunctionality
 {
-    Q_OBJECT
-
 private:
     QQmlApplicationEngine _qmlEngine;
 
 public:
     explicit GUIController(QObject *parent = 0);
-    ~GUIController();
+    virtual ~GUIController();
 
     void init();
