+++ source/common/microkernel_architecture/microkernel_modules/functionalities/AFunctionality.hpp
@@ -5,27 +5,27 @@
 
 class AFunctionality : public QObject
 {  
-  Q_OBJECT
+    Q_OBJECT
 
 private:
-  QThread *_thread;
-  bool _running;
+    QThread *_thread;
+    bool _running;
 
 public:
-  AFunctionality();
-  virtual ~AFunctionality();
-  
-  bool start(bool threaded = true);
-  void stop();
-  bool isRunning() const;
-
+    AFunctionality();
+    virtual ~AFunctionality();
+    
+    bool start(bool threaded = true);
+    void stop();
+    bool isRunning() const;
+			  
 private slots:
-  void _fctStarted();
-  void _fctFinished();
+    void _fctStarted();
+    void _fctFinished();
 
 signals:
-  void started();
-  void stopped();
+    void started();
+    void stopped();
 };
 
 #endif // !AFUNCTIONALITY_HPP_
