+++ source/server/microkernel_architecture/microkernel_modules/functionalities_manager/FunctionalitiesManager.hpp
@@ -1,0 +1,16 @@
+#ifndef FUNCTIONALITIES_MANAGER_HPP_
+# define FUNCTIONALITIES_MANAGER_HPP_
+
+# include "AFunctionalitiesManager.hpp"
+
+class FunctionalitiesManager : private AFunctionalitiesManager
+{
+public:
+    FunctionalitiesManager();
+    virtual ~FunctionalitiesManager();
+
+private:
+    virtual void _setFcts();
+};
+
+#endif // !FUNCTIONALITIES_MANAGER_HPP_
